backend/requirements.txt
fastapi
uvicorn[standard]
celery >= 5.0
python-dotenv
requests # For API calls
google-generativeai
# GitPython is NOT needed for this version